Move testing code to new controller
Cleanup HomeController
